Reorganize routes structure
